requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
plotly==5.18.0
anthropic==0.26.0
httpx==0.27.0
python-dotenv==1.0.1
scipy==1.12.0
scikit-learn==1.4.0
seaborn==0.13.2
xlrd==2.0.1
pyarrow==15.0.0
psutil==5.9.8
